Internal package setup for biomes

Reproducibly classifies occurrence records into biomes using 31 published global biome schemes compiled by Fischer and colleagues (2022) doi:10.1111/geb.13574 , provided as harmonised raster layers at 10x10 km resolution globally. Includes functions to choose the most suitable biome scheme for a dataset by a data-driven ranking, to classify occurrence records, and to tabulate and visualise the result. Works with user-provided occurrences or a taxon name, in which case occurrences are downloaded from GBIF (https://www.gbif.org) and cleaned automatically.

See also

Author

Maintainer: Hans Christian Groß grossha@uni-marburg.de

Authors:

Hans Christian Groß grossha@uni-marburg.de

Alexander Zizka alexander.zizka@biologie.uni-marburg.de [funder]

Anna Walentowitz

Jan-Christopher Fischer