Internal package setup for biomes

Provides raster layers of 31 commonly used biome definitions (from Fischer et al, 2022) at 10x10 km resolutino globally. Furthermore contains two functions to classify species occurrences and species into these biomes and visualize. Works with user-provided occurrences or a species name, in which case occurrences are downloaded from www.gbif.org and cleaned automatically.

Author

Maintainer: Hans Christian Groß grossc@uni-marburg.de

Authors:

Hans Christian Groß grossc@uni-marburg.de

Alexander Zizka alexander.zizka@biologie.uni-marburg.de [funder]